requirements.txt
python-dotenv
aiogram>=3.5
fastapi
uvicorn
pydantic
langchain>=0.2
langchain-openai>=0.1
langchain-community
pymongo